Asteroids, lavender powerup, and some panel changes
Not exactly keen on how the boost and reverse button looks now. Will update when I have a pressed version. Also added the back panel texture and the front panel textures. Can we put mph/s as a grey item below the numbers? It doesn't exactly look great appended to the numbers. Speed and accel labels aren't exactly pretty either inside the black screens.

Forgot how to make spawners pick from an array of 2D textures, so I'll leave that for later.

Saw the lavender modulated powerup and made a special lavender one and replaced it!